[datagen] Move datagenerator.py to top level (see commit message)
This makes it easier to import it from app.py, which is necessary when running
`flask initdb`. We could keep it in a separate Datagen/ directory, but we need
to make Datagen/ its own module then (by adding an __init__.py) in there.
However, datagenerator.py creates the separate Data/ directory when exporting
to csv, and nothing when writing directly to the database.